Move Platform into domain package
