+++ app/controllers/VersionsController.scala
@@ -2,9 +2,9 @@ package controllers
 
 import clients.StateApiImpl
 import com.google.inject.Inject
+import domain.Platform
 import play.api.mvc._
 import repos.CandidatesRepository
-import utils.Platform
 
 import scala.concurrent.ExecutionContext.Implicits.global
