+++ app/controllers/VersionsListController.scala
@@ -3,11 +3,12 @@ package controllers
 import cats.data.OptionT
 import cats.implicits._
 import clients.StateApiImpl
+import domain.Platform
 import ordering.VersionItemOrdering
 import play.api.mvc._
 import rendering.{RowCountCalculator, VersionItemListBuilder, VersionRendering, VersionRow}
 import repos.CandidatesRepository
-import utils.{Platform, VersionListProperties}
+import utils.VersionListProperties
 
 import javax.inject.Inject
 import scala.concurrent.ExecutionContext.Implicits.global
