+++ test/utils/PlatformSpec.scala
@@ -1,5 +1,6 @@
 package utils
 
+import domain.Platform
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
